requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
pydantic
gtts
plotly
torch
transformers
Pillow
pycountry
pandas